+++ README.md
@@ -7,5 +7,6 @@
 |   code challenge 2   |   [link](./challenges/challenge2/challenge2.md)   |
 |   code challenge 3   |   [link](./challenges/challenge3/challenge3.md)   |
 |   code challenge 4   |   [link](./challenges/challenge4/challenge4.md)   |
-|   code challenge 5   |   [link](./challenges/challenge5/README.md)       |
+|   code challenge 5   |   [link](./challenges/link_list/challenge5/linklist/README.md)       |
+|   code challenge 6   |   [link](./challenges/link_list/challenge6/linklist/README.md)       |
