--- challenges/challenge5/linklist/main.py
@@ -1,13 +1,0 @@
-
-from linked_list import Linklist
-if __name__ == "__main__":
-    
-    Linklist1=Linklist()
-    # Linklist1.insert("a")
-    # Linklist1.insert("b")
-    # Linklist1.insert("c")
-    print(Linklist1)
-
-  
-
-    print(Linklist1.includes("b"))
