--- challenges/challenge6/list_insertions/main_insert.py
@@ -1,12 +1,0 @@
-from list_insertions import Linklist2
-if __name__ == "__main__":
-    
-    Linklist1=Linklist2()
-    # Linklist1.insert("a")
-    # Linklist1.insert("b")
-    # Linklist1.insert("c")
-    # print(Linklist1)
-
-  
-
-    # print(Linklist1.includes("b"))
