--- challenges/challenge6/test_insert.py
@@ -1,4 +1,0 @@
-import pytest
-from challenge6.list_insertions.list_insertions import Linklist2
-
-
