+++ challenges/link_list/__pycache__/__init__.cpython-310.pyc
