+++ challenges/link_list/challenge5/linklist/README.md
@@ -1,7 +1,8 @@
 # Challenge Title
 ##  Implementation: Singly Linked Lists
 # Whiteboard Process
-![](../image/chall5.jpg)
+![](../../../image/chall5.jpg)
+
 ## Setup
 req.txt to install the requred
