+++ challenges/link_list/challenge5/linklist/__pycache__/__init__.cpython-310.pyc
