+++ challenges/link_list/challenge5/linklist/__pycache__/node.cpython-310.pyc
