+++ challenges/link_list/challenge6/linklist/README.md
@@ -1,0 +1,17 @@
+# Challenge Title
+##  Implementation:linked-list-insertions
+# Whiteboard Process
+![](../../../image/chall6.jpg)
+## Setup
+req.txt to install the requred
+
+
+### to test code 
+install pytest to run the test
+ ```
+ pytest
+ ```
+ ### to run a code 
+ ```
+ python challenges/challenge5/linklist/main.py
+ ```
